add basic input section component
