+++ src/app/components/sidebar/object-inpsector/basic-section/basic-input-section.component.scss
@@ -1,0 +1,8 @@
+.section-text {
+    margin-bottom: 10px;
+}
+
+.align-right {
+    flex: 1;
+    text-align: right;
+}
