+++ src/app/components/sidebar/object-inpsector/input-inspector.component.scss
@@ -9,11 +9,15 @@
     border-left: 1px solid $border-color;
 }
 
-.edit-text {
+.input {
     font-size: 1.2em;
 }
 
 .info-icon {
     margin-left: 10px;
     cursor: pointer;
+}
+
+.input-text {
+    font-size: 1.2rem;
 }
